requirements.txt
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.245
numpy==1.23.5
scipy==1.15.3
opencv-python==4.9.0.80
PyYAML==6.0.3
pandas==2.3.3
matplotlib==3.10.8
lapx==0.9.4
filterpy==1.4.5
loguru==0.7.3
tqdm==4.67.3
requests==2.32.5
gdown==5.2.1
filelock==3.20.1
click==8.3.1
packaging==25.0
Pillow==12.0.0
ftfy==6.3.1
regex==2024.11.6
yacs==0.1.8
